오버플로우 수정

+++ BeetMall/src/main/java/com/beetmall/sshj/custom/vo/MyBuyListPageVO.java
@@ -10,7 +10,7 @@ public class MyBuyListPageVO {
 	private String orderstatus;
 	private int productnum;
 	private long invoice;
-	private int ordercnt;
+	private long ordercnt;
 	 
 	public String getOrderdate() {
 		return orderdate;
@@ -66,10 +66,10 @@ public class MyBuyListPageVO {
 	public void setInvoice(long invoice) {
 		this.invoice = invoice;
 	}
-	public int getOrdercnt() {
+	public long getOrdercnt() {
 		return ordercnt;
 	}
-	public void setOrdercnt(int ordercnt) {
+	public void setOrdercnt(long ordercnt) {
 		this.ordercnt = ordercnt;
 	}
